Refactor: Stabilize config and shared type entrypoints
