+++ src/shared/types/index.ts
@@ -1,0 +1,574 @@
+interface Common {
+  id: string;
+  name: string;
+  position?: string;
+  games: number;
+  goals: number;
+  assists: number;
+  points: number;
+  penalties: number;
+  ppp: number;
+  shp: number;
+  score: number;
+  scoreAdjustedByGames: number;
+  scores?: Record<string, number>;
+  scoreByPosition?: number;
+  scoreByPositionAdjustedByGames?: number;
+  scoresByPosition?: Record<string, number>;
+}
+
+export interface Player extends Common {
+  plusMinus: number;
+  shots: number;
+  hits: number;
+  blocks: number;
+}
+
+export interface Goalie extends Common {
+  wins: number;
+  saves: number;
+  shutouts: number;
+  gaa?: string;
+  savePercent?: string;
+}
+
+export interface PlayerWithSeason extends Player {
+  season: number;
+}
+
+// Single-season snapshot used inside CombinedPlayer.seasons.
+// Includes per-season score, scoreAdjustedByGames and scores, but omits name
+// (name is available at the CombinedPlayer root level).
+export type PlayerSeasonData = Omit<PlayerWithSeason, "name">;
+
+export interface CombinedPlayer extends Player {
+  seasons: PlayerSeasonData[];
+}
+
+export interface GoalieWithSeason extends Goalie {
+  season: number;
+}
+
+// Single-season snapshot used inside CombinedGoalie.seasons.
+// Includes per-season score, scoreAdjustedByGames and scores, but omits name
+// (name is available at the CombinedGoalie root level).
+export type GoalieSeasonData = Omit<GoalieWithSeason, "name">;
+
+export interface CombinedGoalie extends Goalie {
+  seasons: GoalieSeasonData[];
+}
+
+export type CountSplit = {
+  owned: number;
+  played: number;
+};
+
+export type CareerSummaryTeam = {
+  teamId: string;
+  teamName: string;
+  seasonCount: CountSplit;
+  firstSeason: number;
+  lastSeason: number;
+};
+
+export type CareerSummary = {
+  firstSeason: number;
+  lastSeason: number;
+  seasonCount: CountSplit;
+  teamCount: CountSplit;
+  teams: CareerSummaryTeam[];
+};
+
+export type CareerPlayerSeasonRow = {
+  season: number;
+  reportType: CsvReport;
+  teamId: string;
+  teamName: string;
+  position: string;
+  games: number;
+  goals: number;
+  assists: number;
+  points: number;
+  plusMinus: number;
+  penalties: number;
+  shots: number;
+  ppp: number;
+  shp: number;
+  hits: number;
+  blocks: number;
+};
+
+export type CareerGoalieSeasonRow = {
+  season: number;
+  reportType: CsvReport;
+  teamId: string;
+  teamName: string;
+  games: number;
+  wins: number;
+  saves: number;
+  shutouts: number;
+  goals: number;
+  assists: number;
+  points: number;
+  penalties: number;
+  ppp: number;
+  shp: number;
+  gaa?: string;
+  savePercent?: string;
+};
+
+export type CareerPlayerTeamTotals = {
+  teamId: string;
+  teamName: string;
+  seasonCount: CountSplit;
+  games: number;
+  goals: number;
+  assists: number;
+  points: number;
+  plusMinus: number;
+  penalties: number;
+  shots: number;
+  ppp: number;
+  shp: number;
+  hits: number;
+  blocks: number;
+};
+
+export type CareerGoalieTeamTotals = {
+  teamId: string;
+  teamName: string;
+  seasonCount: CountSplit;
+  games: number;
+  wins: number;
+  saves: number;
+  shutouts: number;
+  goals: number;
+  assists: number;
+  points: number;
+  penalties: number;
+  ppp: number;
+  shp: number;
+};
+
+export type CareerPlayerTotals = {
+  seasonCount: CountSplit;
+  teamCount: CountSplit;
+  teams: CareerPlayerTeamTotals[];
+  games: number;
+  goals: number;
+  assists: number;
+  points: number;
+  plusMinus: number;
+  penalties: number;
+  shots: number;
+  ppp: number;
+  shp: number;
+  hits: number;
+  blocks: number;
+};
+
+export type CareerGoalieTotals = {
+  seasonCount: CountSplit;
+  teamCount: CountSplit;
+  teams: CareerGoalieTeamTotals[];
+  games: number;
+  wins: number;
+  saves: number;
+  shutouts: number;
+  goals: number;
+  assists: number;
+  points: number;
+  penalties: number;
+  ppp: number;
+  shp: number;
+};
+
+export type CareerPlayerResponse = {
+  id: string;
+  name: string;
+  position: string;
+  summary: CareerSummary;
+  totals: {
+    career: CareerPlayerTotals;
+    regular: CareerPlayerTotals;
+    playoffs: CareerPlayerTotals;
+  };
+  seasons: CareerPlayerSeasonRow[];
+};
+
+export type CareerGoalieResponse = {
+  id: string;
+  name: string;
+  summary: CareerSummary;
+  totals: {
+    career: CareerGoalieTotals;
+    regular: CareerGoalieTotals;
+    playoffs: CareerGoalieTotals;
+  };
+  seasons: CareerGoalieSeasonRow[];
+};
+
+export type CareerPlayerListItem = {
+  id: string;
+  name: string;
+  position: string;
+  firstSeason: number;
+  lastSeason: number;
+  seasonsOwned: number;
+  seasonsPlayedRegular: number;
+  seasonsPlayedPlayoffs: number;
+  teamsOwned: number;
+  teamsPlayedRegular: number;
+  teamsPlayedPlayoffs: number;
+  regularGames: number;
+  playoffGames: number;
+};
+
+export type CareerGoalieListItem = {
+  id: string;
+  name: string;
+  firstSeason: number;
+  lastSeason: number;
+  seasonsOwned: number;
+  seasonsPlayedRegular: number;
+  seasonsPlayedPlayoffs: number;
+  teamsOwned: number;
+  teamsPlayedRegular: number;
+  teamsPlayedPlayoffs: number;
+  regularGames: number;
+  playoffGames: number;
+};
+
+export type CareerHighlightType =
+  | "most-teams-played"
+  | "most-teams-owned"
+  | "same-team-seasons-played"
+  | "same-team-seasons-owned"
+  | "most-stanley-cups"
+  | "reunion-king"
+  | "stash-king"
+  | "regular-grinder-without-playoffs"
+  | "most-trades"
+  | "most-claims"
+  | "most-drops";
+
+export type CareerHighlightTeam = {
+  id: string;
+  name: string;
+};
+
+export type CareerTransactionHighlightTeam = CareerHighlightTeam & {
+  count: number;
+};
+
+export type CareerTeamCountHighlightItem = {
+  id: string;
+  name: string;
+  position: string;
+  teamCount: number;
+  teams: CareerHighlightTeam[];
+};
+
+export type CareerSameTeamHighlightItem = {
+  id: string;
+  name: string;
+  position: string;
+  seasonCount: number;
+  team: CareerHighlightTeam;
+};
+
+export type CareerStanleyCupHighlightCup = {
+  season: number;
+  team: CareerHighlightTeam;
+};
+
+export type CareerStanleyCupHighlightItem = {
+  id: string;
+  name: string;
+  position: string;
+  cupCount: number;
+  cups: CareerStanleyCupHighlightCup[];
+};
+
+export type CareerReunionType = "claim" | "trade";
+
+export type CareerReunionHighlightReunion = {
+  date: string;
+  type: CareerReunionType;
+};
+
+export type CareerReunionHighlightItem = {
+  id: string;
+  name: string;
+  position: string;
+  reunionCount: number;
+  team: CareerHighlightTeam;
+  reunions: CareerReunionHighlightReunion[];
+};
+
+export type CareerStashHighlightItem = {
+  id: string;
+  name: string;
+  position: string;
+  seasonCount: number;
+  team: CareerHighlightTeam;
+};
+
+export type CareerRegularGrinderHighlightItem = {
+  id: string;
+  name: string;
+  position: string;
+  regularGames: number;
+  teams: CareerHighlightTeam[];
+};
+
+export type CareerTransactionHighlightItem = {
+  id: string;
+  name: string;
+  position: string;
+  transactionCount: number;
+  teams: CareerTransactionHighlightTeam[];
+};
+
+export type CareerTeamCountHighlightPage = {
+  type: "most-teams-played" | "most-teams-owned";
+  minAllowed: number;
+  skip: number;
+  take: number;
+  total: number;
+  items: CareerTeamCountHighlightItem[];
+};
+
+export type CareerSameTeamHighlightPage = {
+  type: "same-team-seasons-played" | "same-team-seasons-owned";
+  minAllowed: number;
+  skip: number;
+  take: number;
+  total: number;
+  items: CareerSameTeamHighlightItem[];
+};
+
+export type CareerStanleyCupHighlightPage = {
+  type: "most-stanley-cups";
+  minAllowed: number;
+  skip: number;
+  take: number;
+  total: number;
+  items: CareerStanleyCupHighlightItem[];
+};
+
+export type CareerReunionHighlightPage = {
+  type: "reunion-king";
+  minAllowed: number;
+  skip: number;
+  take: number;
+  total: number;
+  items: CareerReunionHighlightItem[];
+};
+
+export type CareerStashHighlightPage = {
+  type: "stash-king";
+  minAllowed: number;
+  skip: number;
+  take: number;
+  total: number;
+  items: CareerStashHighlightItem[];
+};
+
+export type CareerRegularGrinderHighlightPage = {
+  type: "regular-grinder-without-playoffs";
+  minAllowed: number;
+  skip: number;
+  take: number;
+  total: number;
+  items: CareerRegularGrinderHighlightItem[];
+};
+
+export type CareerTransactionHighlightPage = {
+  type: "most-trades" | "most-claims" | "most-drops";
+  minAllowed: number;
+  skip: number;
+  take: number;
+  total: number;
+  items: CareerTransactionHighlightItem[];
+};
+
+export type CareerHighlightsPage =
+  | CareerTeamCountHighlightPage
+  | CareerSameTeamHighlightPage
+  | CareerStanleyCupHighlightPage
+  | CareerReunionHighlightPage
+  | CareerStashHighlightPage
+  | CareerRegularGrinderHighlightPage
+  | CareerTransactionHighlightPage;
+
+export type PlayerFields =
+  | "name"
+  | "games"
+  | "goals"
+  | "assists"
+  | "points"
+  | "plusMinus"
+  | "penalties"
+  | "shots"
+  | "ppp"
+  | "shp"
+  | "hits"
+  | "blocks";
+
+export type GoalieFields =
+  | "name"
+  | "games"
+  | "wins"
+  | "saves"
+  | "shutouts"
+  | "goals"
+  | "assists"
+  | "points"
+  | "penalties"
+  | "ppp"
+  | "shp";
+
+export interface RawData {
+  Skaters: string;
+  season: number;
+  field2: string;
+  field3: string;
+  field4: string;
+  field5: string;
+  field6: string;
+  field7: string;
+  field8: string;
+  field9: string;
+  field10: string;
+  field11: string;
+  field12: string;
+  field13: string;
+  field14: string;
+  field15: string;
+  field16: string;
+  field17: string;
+  field18: string;
+  field19?: string;
+  field20?: string;
+}
+
+export type PlayerScoreField =
+  | "goals"
+  | "assists"
+  | "points"
+  | "plusMinus"
+  | "penalties"
+  | "shots"
+  | "ppp"
+  | "shp"
+  | "hits"
+  | "blocks";
+
+export type GoalieScoreField = "wins" | "saves" | "shutouts";
+
+export type GoalieOptionalScoreField = "gaa" | "savePercent";
+
+export type PlayerScoreWeights = Record<PlayerScoreField, number>;
+export type GoalieScoreWeights = Record<
+  GoalieScoreField | GoalieOptionalScoreField,
+  number
+>;
+
+export type CsvReport = "regular" | "playoffs";
+
+// API-level reportType. `both` is virtual and represents regular+playoffs merged.
+export type Report = CsvReport | "both";
+
+export type QueryParams = {
+  reportType: Report;
+  season?: number;
+  startFrom?: number;
+};
+
+export type Team = {
+  id: string;
+  name: string;
+  presentName: string;
+  nameAliases?: string[];
+  // First season year in the YYYY-YYYY+1 format used by imports (e.g. 2017 => 2017-2018).
+  // Useful for expansion/relocation where a team doesn't exist in older seasons.
+  firstSeason?: number;
+};
+
+export type PlayoffLeaderboardEntry = {
+  teamId: string;
+  teamName: string;
+  appearances: number;
+  championships: number;
+  finals: number;
+  conferenceFinals: number;
+  secondRound: number;
+  firstRound: number;
+  seasons: PlayoffLeaderboardSeason[];
+  tieRank: boolean;
+};
+
+export type RegularLeaderboardSeason = {
+  season: number;
+  regularTrophy: boolean;
+  wins: number;
+  losses: number;
+  ties: number;
+  points: number;
+  divWins: number;
+  divLosses: number;
+  divTies: number;
+  winPercent: number;
+  divWinPercent: number;
+  pointsPercent: number;
+};
+
+export type PlayoffRoundKey =
+  | "championship"
+  | "final"
+  | "conferenceFinal"
+  | "secondRound"
+  | "firstRound"
+  | "notQualified";
+
+export type PlayoffLeaderboardSeason = {
+  season: number;
+  round: number;
+  key: PlayoffRoundKey;
+};
+
+export type RegularLeaderboardEntry = {
+  teamId: string;
+  teamName: string;
+  wins: number;
+  losses: number;
+  ties: number;
+  points: number;
+  divWins: number;
+  divLosses: number;
+  divTies: number;
+  winPercent: number;
+  divWinPercent: number;
+  pointsPercent: number;
+  regularTrophies: number;
+  seasons: RegularLeaderboardSeason[];
+  tieRank: boolean;
+};
+
+export type TransactionLeaderboardSeason = {
+  season: number;
+  claims: number;
+  drops: number;
+  trades: number;
+};
+
+export type TransactionLeaderboardEntry = {
+  teamId: string;
+  teamName: string;
+  claims: number;
+  drops: number;
+  trades: number;
+  seasons: TransactionLeaderboardSeason[];
+  tieRank: boolean;
+};
